ca/pyproject.toml
[tool.poetry]
name = "my_agent"
version = "0.1.0"
description = "Example LangGraph project for deployment to LangGraph Cloud"
authors = [
    "langchain-ai"
]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langgraph = "^0.2.0"
langchain_anthropic = "^0.1.0"
langchain_core = "^0.2.0"
langchain_openai = "^0.1.0"
tavily-python = "^0.3.0"
langchain_community = "^0.2.0"
nest-asyncio = "^1.6.0"
uvicorn = "^0.30.6"
langserve = "^0.2.3"
fastapi = "^0.114.0"
sse-starlette = "^2.1.3"
langgraph-checkpoint-sqlite = "^1.0.3"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"